jenkins/project/app/django/requirements.txt
asgiref==3.3.4
Django==3.2
django-appconf==1.0.4
django-compressor==2.4.1
django-crispy-forms==1.11.2
django-libsass==0.8
django-translations==1.2.0
gunicorn==20.1.0
libsass==0.20.1
Pillow==7.0.0
psycopg2-binary==2.8.5
pytz==2021.1
rcssmin==1.0.6
rjsmin==1.1.0
selenium==3.141.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.4
